test: set up and tear down for e2e testing
