+++ features/export.feature
@@ -1,7 +1,8 @@
 Feature: cli-kintone export command
 
   Scenario: CliKintoneTest-81 Should return the record contents in CSV format
-    Given The app "$$TEST_KINTONE_APP_ID_FOR_EXPORT" has some records as below:
+    Given The app "$$TEST_KINTONE_APP_ID_FOR_EXPORT" with "$$TEST_KINTONE_API_TOKEN_FOR_DELETE" has no records
+    And The app "$$TEST_KINTONE_APP_ID_FOR_EXPORT" has some records as below:
       | Text   | Number |
       | Alice  | 10     |
       | Bob    | 20     |
