+++ features/step_definitions/common.ts
@@ -1,5 +1,9 @@
 import * as assert from "assert";
-import { execCliKintoneSync, replaceTokenWithEnvVars } from "../ultils/helper";
+import {
+  createCsvFile,
+  execCliKintoneSync,
+  replaceTokenWithEnvVars,
+} from "../ultils/helper";
 import { Given, When, Then } from "../ultils/world";
 
 Given(
@@ -10,6 +14,30 @@ Given(
       throw new Error(`The env variable is missing: ${srcKey}`);
     }
     this.env = { [destKey]: value, ...this.env };
+  },
+);
+
+Given(
+  "The app {string} with {string} has no records",
+  function (appId: string, apiToken: string) {
+    const command = `record delete --app ${appId} --base-url $$TEST_KINTONE_BASE_URL --api-token ${apiToken} --yes`;
+    const response = execCliKintoneSync(replaceTokenWithEnvVars(command));
+    if (response.status !== 0) {
+      throw new Error(`Resetting app failed. Error: \n${response.stderr}`);
+    }
+  },
+);
+
+Given(
+  "The app {string} has some records as below:",
+  async function (appId, table) {
+    const tempFilePath = await createCsvFile(table.raw());
+    const command = `record import --file-path ${tempFilePath} --app ${appId} --base-url $$TEST_KINTONE_BASE_URL --username $$TEST_KINTONE_USERNAME --password $$TEST_KINTONE_PASSWORD`;
+
+    const response = execCliKintoneSync(replaceTokenWithEnvVars(command));
+    if (response.status !== 0) {
+      throw new Error(`Importing CSV failed. Error: \n${response.stderr}`);
+    }
   },
 );
